transcription-analysis/requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
groq>=0.4.1
python-dotenv>=1.0.0
pytube>=15.0.0
youtube-transcript-api>=0.6.2 